Move block entity capability interfaces and lookups into new packages
