+++ src/main/java/com/logistics/LogisticsCore.java
@@ -3,10 +3,10 @@ package com.logistics;
 import com.logistics.core.bootstrap.DomainBootstrap;
 import com.logistics.core.item.ProbeItem;
 import com.logistics.core.item.WrenchItem;
-import com.logistics.core.lib.block.storage.EnergyStorageAccess;
-import com.logistics.core.lib.block.storage.FluidStorageAccess;
-import com.logistics.core.lib.block.storage.ItemStorageAccess;
-import com.logistics.core.lib.block.storage.PipeConnectionAccess;
+import com.logistics.core.lib.block.lookup.EnergyStorageAccess;
+import com.logistics.core.lib.block.lookup.FluidStorageAccess;
+import com.logistics.core.lib.block.lookup.ItemStorageAccess;
+import com.logistics.core.lib.block.lookup.PipeConnectionAccess;
 import com.logistics.core.loot.ChestLootModifier;
 import com.logistics.core.marker.MarkerBlock;
 import com.logistics.core.marker.MarkerBlockEntity;
