+++ src/main/java/com/logistics/automation/laserquarry/entity/LaserQuarryBlockEntity.java
@@ -9,8 +9,8 @@ import com.logistics.automation.laserquarry.LaserQuarryFrameBlock;
 import com.logistics.automation.render.ClientRenderCacheHooks;
 import com.logistics.core.lib.BaseBlockEntity;
 import com.logistics.core.lib.energy.EnergyComponent;
-import com.logistics.core.lib.entity.HasEnergyStorage;
-import com.logistics.core.lib.pipe.PipeConnection;
+import com.logistics.core.lib.block.capability.HasEnergyStorage;
+import com.logistics.core.lib.block.capability.PipeConnection;
 import com.logistics.core.lib.storage.NbtCompat;
 import com.logistics.core.lib.support.ProbeResult;
 import java.util.ArrayList;
