+++ src/main/java/com/logistics/core/lib/block/lookup/ItemStorageAccess.java
@@ -1,6 +1,6 @@
-package com.logistics.core.lib.block.storage;
+package com.logistics.core.lib.block.lookup;
 
-import com.logistics.core.lib.entity.HasItemStorage;
+import com.logistics.core.lib.block.capability.HasItemStorage;
 import net.fabricmc.fabric.api.transfer.v1.item.ItemStorage;
 
 /**
