+++ src/main/java/com/logistics/core/lib/power/AbstractEngineBlockEntity.java
@@ -1,7 +1,7 @@
 package com.logistics.core.lib.power;
 
 import com.logistics.core.lib.BaseBlockEntity;
-import com.logistics.core.lib.entity.HasEnergyStorage;
+import com.logistics.core.lib.block.capability.HasEnergyStorage;
 import com.logistics.core.lib.storage.NbtCompat;
 import com.logistics.core.lib.support.ProbeResult;
 import java.util.Locale;
