+++ src/main/java/com/logistics/pipe/block/entity/PipeBlockEntity.java
@@ -4,10 +4,10 @@ import com.logistics.LogisticsMod;
 import com.logistics.LogisticsPipe;
 import com.logistics.core.lib.BaseBlockEntity;
 import com.logistics.core.lib.energy.EnergyComponent;
-import com.logistics.core.lib.entity.HasEnergyStorage;
-import com.logistics.core.lib.entity.HasItemStorage;
+import com.logistics.core.lib.block.capability.HasEnergyStorage;
+import com.logistics.core.lib.block.capability.HasItemStorage;
 import com.logistics.core.lib.storage.NbtCompat;
-import com.logistics.core.lib.pipe.PipeConnection;
+import com.logistics.core.lib.block.capability.PipeConnection;
 import com.logistics.core.lib.power.AcceptsLowTierEnergy;
 import com.logistics.pipe.Pipe;
 import com.logistics.pipe.PipeContext;
