+++ src/main/java/com/logistics/power/block/entity/CreativeSinkBlockEntity.java
@@ -2,7 +2,7 @@ package com.logistics.power.block.entity;
 
 import com.logistics.LogisticsPower;
 import com.logistics.core.lib.BaseBlockEntity;
-import com.logistics.core.lib.entity.HasEnergyStorage;
+import com.logistics.core.lib.block.capability.HasEnergyStorage;
 import com.logistics.core.lib.power.AcceptsLowTierEnergy;
 import com.logistics.core.lib.storage.NbtCompat;
 import com.logistics.core.lib.support.ProbeResult;
